SnowflakeOparations/requirements.txt
snowflake-connector-python>=3.0.0
pandas>=1.5.0
openpyxl>=3.0.0
